+++ test/mitmproxy/data/addonscripts/duplicate_flow.py
@@ -2,5 +2,5 @@ from mitmproxy import ctx
 
 
 def request(flow):
-    f = ctx.master.duplicate_flow(flow)
+    f = ctx.master.state.duplicate_flow(flow)
     ctx.master.replay_request(f, block=True)
