+++ test/mitmproxy/test_flow.py
@@ -452,17 +452,6 @@ class TestFlowMaster:
         f.live = True
         tutils.raises("live", fm.replay_request, f)
 
-    def test_duplicate_flow(self):
-        s = flow.State()
-        fm = flow.FlowMaster(None, None, s)
-        f = tutils.tflow(resp=True)
-        fm.load_flow(f)
-        assert s.flow_count() == 1
-        f2 = fm.duplicate_flow(f)
-        assert f2.response
-        assert s.flow_count() == 2
-        assert s.index(f2) == 1
-
     def test_create_flow(self):
         s = flow.State()
         fm = flow.FlowMaster(None, None, s)
